% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{check_response_inversion}
\alias{check_response_inversion}
\title{Check whether response codes should be flipped (not yet implemented)}
\usage{
check_response_inversion(...)
}
\arguments{
\item{...}{Any arguments; ignored.}
}
\value{
Not implemented. Currently calls \code{\link[cli:cli_abort]{cli::cli_abort()}}.
}
\description{
Planned behaviour: compute the CI with the original response codes and
with the codes flipped, compare the information value of the two, and
flag the data as potentially inverted if the flipped version has a
substantially higher infoVal.
}
