% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcdiag_result.R
\name{is_rcdiag_result}
\alias{is_rcdiag_result}
\title{Test whether an object is an \code{rcdiag_result}}
\usage{
is_rcdiag_result(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"rcdiag_result"}, otherwise \code{FALSE}.
}
\description{
Test whether an object is an \code{rcdiag_result}
}
