% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcicrdiagnostics-package.R
\docType{package}
\name{rcicrdiagnostics-package}
\alias{rcicrdiagnostics}
\alias{rcicrdiagnostics-package}
\title{rcicrdiagnostics: Data-Quality Diagnostics for Reverse Correlation Experiments}
\description{
A systematic battery of data-quality checks for reverse correlation experiments. Supports both the standard two-image forced-choice (2IFC) pipeline via the 'rcicr' package and the Brief-RC pipeline (Schmitz et al., 2024) that uses direct noise matrix multiplication. Designed to be run before computing classification images or information values so that silent data-processing errors are caught early.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/olivethree/rcicrdiagnostics}
  \item \url{https://olivethree.github.io/rcicrdiagnostics/}
  \item Report bugs at \url{https://github.com/olivethree/rcicrdiagnostics/issues}
}

}
\author{
\strong{Maintainer}: Manuel Oliveira \email{manuel.olivethree@gmail.com}

}
\keyword{internal}
