% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{check_version_compat}
\alias{check_version_compat}
\title{Check rcicr version compatibility (not yet implemented)}
\usage{
check_version_compat(...)
}
\arguments{
\item{...}{Any arguments; ignored.}
}
\value{
Not implemented.
}
\description{
Planned behaviour: verify that the rcicr version used to generate the
rdata file is compatible with the installed rcicr version used to
reconstruct CIs.
}
